+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - (Feature) Adds `<<` and `>>` to move tabs.
 - (Bugfix) `<shift>+f` opens links in new tab.
+- (Bugfix) Fixes search with `/`.
 - (Misc) Shows notifications less intrusive on the lower right.
 - (Misc) Shows labels for links in lower case instead of upper case.